BinomialSeparate class

This class represents a Truncated Gaussian model using the CPP (Conditional Power Prior) approach with power parameter set to 0. It inherits from the BinomialCPP class.

Super classes

RBExT::Model -> RBExT::MCMCModel -> BinomialSeparate

Public fields

method

Method name

Methods

Public methods

BinomialSeparate$new()

BinomialSeparate$prepare_data()

BinomialSeparate$sample_prior()

BinomialSeparate$prior_pdf()

BinomialSeparate$prior_cdf()

BinomialSeparate$clone()

Inherited methods

RBExT::Model$create()

RBExT::Model$empirical_bayes_update()

RBExT::Model$estimate_bayesian_operating_characteristics()

RBExT::Model$estimate_frequentist_operating_characteristics()

RBExT::Model$plot_pdfs()

RBExT::Model$plot_posterior_pdf()

RBExT::Model$plot_prior_pdf()

RBExT::Model$posterior_mean()

RBExT::Model$posterior_moments()

RBExT::Model$posterior_quantile()

RBExT::Model$posterior_to_RBesT()

RBExT::Model$prior_ESS()

RBExT::Model$prior_to_RBesT()

RBExT::Model$prior_treatment_benefit()

RBExT::Model$simulation_for_given_treatment_effect()

RBExT::Model$test_decision()

RBExT::MCMCModel$check_data()

RBExT::MCMCModel$check_mcmc_config()

RBExT::MCMCModel$compute_posterior_parameters()

RBExT::MCMCModel$credible_interval()

RBExT::MCMCModel$draw_mcmc_prior()

RBExT::MCMCModel$inference()

RBExT::MCMCModel$posterior_cdf()

RBExT::MCMCModel$posterior_median()

RBExT::MCMCModel$posterior_pdf()

RBExT::MCMCModel$sample_posterior()

Method new()

Usage

BinomialSeparate$new(prior, mcmc_config)

Arguments

prior

Prior

mcmc_config

MCMC configuration

Returns

A Model object.

Method prepare_data()

Prepare the data for the Stan model

Usage

BinomialSeparate$prepare_data(target_data)

Arguments

target_data

The target data for the analysis.

Returns

List containing the input to the Stan model

Method sample_prior()

Sample from the prior

Usage

BinomialSeparate$sample_prior(n_samples)

Arguments

n_samples

Number of samples from the prior

Method prior_pdf()

Prior PDF

Usage

BinomialSeparate$prior_pdf(target_treatment_effect)

Arguments

target_treatment_effect

Point at which to evaluate the prior PDF

Method prior_cdf()

Prior CDF

Usage

BinomialSeparate$prior_cdf(target_treatment_effect)

Arguments

target_treatment_effect

Point at which to evaluate the prior CDF

Method clone()

The objects of this class are cloneable with this method.

Usage

BinomialSeparate$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.