PoolGaussian class

This class represents a model with pooled Gaussian components, inheriting from StaticBorrowingGaussian. The power parameter is set to 1.

Super classes

RBExT::Model -> RBExT::ConjugateGaussian -> RBExT::StaticBorrowingGaussian -> PoolGaussian

Public fields

method

Method name

Methods

Public methods

PoolGaussian$new()

PoolGaussian$clone()

Inherited methods

RBExT::Model$create()

RBExT::Model$empirical_bayes_update()

RBExT::Model$estimate_bayesian_operating_characteristics()

RBExT::Model$estimate_frequentist_operating_characteristics()

RBExT::Model$inference()

RBExT::Model$plot_pdfs()

RBExT::Model$plot_posterior_pdf()

RBExT::Model$plot_prior_pdf()

RBExT::Model$posterior_quantile()

RBExT::Model$prior_ESS()

RBExT::Model$prior_treatment_benefit()

RBExT::Model$simulation_for_given_treatment_effect()

RBExT::Model$test_decision()

RBExT::ConjugateGaussian$credible_interval()

RBExT::ConjugateGaussian$posterior_cdf()

RBExT::ConjugateGaussian$posterior_mean()

RBExT::ConjugateGaussian$posterior_median()

RBExT::ConjugateGaussian$posterior_moments()

RBExT::ConjugateGaussian$posterior_pdf()

RBExT::ConjugateGaussian$posterior_to_RBesT()

RBExT::ConjugateGaussian$posterior_variance()

RBExT::ConjugateGaussian$prior_cdf()

RBExT::ConjugateGaussian$prior_pdf()

RBExT::ConjugateGaussian$prior_to_RBesT()

RBExT::ConjugateGaussian$sample_posterior()

RBExT::ConjugateGaussian$sample_prior()

Method new()

Usage

PoolGaussian$new(prior)

Arguments

prior

Prior

Returns

A Model object.

Method clone()

The objects of this class are cloneable with this method.

Usage

PoolGaussian$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

NA
#> [1] NA