TestThenPoolEquivalence class

This class extends the TestThenPool class to implement the Test-Then-Pool framework for equivalence trials.

Details

The TestThenPoolEquivalence class provides methods for testing and inference in equivalence trials using the Test-Then-Pool framework.

Super classes

RBExT::Model -> RBExT::TestThenPool -> TestThenPoolEquivalence

Public fields

significance_level

Significance level for the test

equivalence_margin

Equivalence margin for the test

method

Method name

Methods

Public methods

TestThenPoolEquivalence$new()

TestThenPoolEquivalence$test_pvalue()

TestThenPoolEquivalence$test()

TestThenPoolEquivalence$clone()

Inherited methods

RBExT::Model$create()

RBExT::Model$estimate_bayesian_operating_characteristics()

RBExT::Model$estimate_frequentist_operating_characteristics()

RBExT::Model$plot_pdfs()

RBExT::Model$plot_posterior_pdf()

RBExT::Model$plot_prior_pdf()

RBExT::Model$posterior_mean()

RBExT::Model$posterior_moments()

RBExT::Model$posterior_quantile()

RBExT::Model$prior_ESS()

RBExT::Model$prior_treatment_benefit()

RBExT::Model$simulation_for_given_treatment_effect()

RBExT::TestThenPool$credible_interval()

RBExT::TestThenPool$empirical_bayes_update()

RBExT::TestThenPool$inference()

RBExT::TestThenPool$plot_test_pvalue_vs_drift()

RBExT::TestThenPool$plot_test_vs_drift()

RBExT::TestThenPool$posterior_cdf()

RBExT::TestThenPool$posterior_median()

RBExT::TestThenPool$posterior_pdf()

RBExT::TestThenPool$posterior_to_RBesT()

RBExT::TestThenPool$prior_cdf()

RBExT::TestThenPool$prior_pdf()

RBExT::TestThenPool$prior_to_RBesT()

RBExT::TestThenPool$sample_posterior()

RBExT::TestThenPool$sample_prior()

RBExT::TestThenPool$test_decision()

Method new()

Initializes a TestThenPoolEquivalence object.

Usage

TestThenPoolEquivalence$new(prior, mcmc_config = NULL)

Arguments

prior

The prior distribution for the treatment effect.

mcmc_config

Configuration for the MCMC sampling.

Method test_pvalue()

Performs the test in the Test-Then-Pool framework for equivalence.

Usage

TestThenPoolEquivalence$test_pvalue(target_data, test_type = "t-test")

Arguments

target_data

The data from the target study.

test_type

Test type

Method test()

Usage

TestThenPoolEquivalence$test(target_data, test_type = "t-test")

Arguments

target_data

Target study data

test_type

Frequentist test

Method clone()

The objects of this class are cloneable with this method.

Usage

TestThenPoolEquivalence$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.